requirements.txt
dominate==2.9.0
dotmap==1.3.30
einops==0.7.0
lib==4.0.0
munch==2.5.0
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1
packaging==23.2
Pillow==10.1.0
PyYAML==6.0.1
scipy==1.11.4
torch
torchvision
tqdm==4.66.1
visdom==0.2.4
yacs==0.1.8
